# reduce variables when functions have indirect calls