--- a/lib/compress/reduce-vars.js
+++ b/lib/compress/reduce-vars.js
@@ -509,15 +509,26 @@ function handle_defined_after_hoist(parent) {
         ) return true;
     });
 
+    // If a function `A` uses an outside variable `B`, calling `A` before `B` is defined
+    // will either cause a ReferenceError (if `B` is declared with `let` or `const`) or
+    // get an `undefined` (if `B` is declared with `var`).
+    //
+    // We want to inline `B` into `A` if `B` is a constant and only used in `A`, but we can't
+    // do that if the above situation happens, as it would change the behavior.
+    //
+    // This function checks for such use-before-define situations, including the case where
+    // `A` is indirectly called by another function `C` that is called before `B` is defined.
+
+    // `defun` id to array of `defun` that use the function
+    const defun_dependents_map = new Map();
+    // `defun` id to array of enclosing `def` that are used by the function
+    const defun_enclosing_defs_map = new Map();
+    // all symbol ids that will be tracked for read/write
     const symbols_of_interest = new Set();
-    const defuns_of_interest = new Set();
-    const potential_conflicts = [];
 
     for (const defun of defuns) {
         const fname_def = defun.name.definition();
-        const found_self_ref_in_other_defuns = defuns.some(
-            d => d !== defun && d.enclosed.indexOf(fname_def) !== -1
-        );
+        const enclosing_defs = [];
 
         for (const def of defun.enclosed) {
             if (
@@ -528,93 +539,144 @@ function handle_defined_after_hoist(parent) {
                 continue;
             }
 
-            // defun is hoisted, so always safe
+            // found a reference to another function
             if (
                 def.assignments === 0
                 && def.orig.length === 1
                 && def.orig[0] instanceof AST_SymbolDefun
             ) {
+                if (!defun_dependents_map.has(def.id)) {
+                    defun_dependents_map.set(def.id, []);
+                }
+                defun_dependents_map.get(def.id).push(fname_def.id);
                 continue;
             }
 
-            if (found_self_ref_in_other_defuns) {
-                def.fixed = false;
-                continue;
-            }
-
-            // for the slower checks below this loop
-            potential_conflicts.push({ defun, def, fname_def });
+            enclosing_defs.push(def);
             symbols_of_interest.add(def.id);
-            symbols_of_interest.add(fname_def.id);
-            defuns_of_interest.add(defun);
+        }
+
+        if (enclosing_defs.length) {
+            defun_enclosing_defs_map.set(fname_def.id, enclosing_defs);
         }
     }
 
-    // linearize all symbols, and locate defs that are read after the defun
-    if (potential_conflicts.length) {
-        // All "symbols of interest", that is, defuns or defs, that we found.
-        // These are placed in order so we can check which is after which.
-        const found_symbols = [];
-        // Indices of `found_symbols` which are writes
-        const found_symbol_writes = new Set();
-        // Defun ranges are recorded because we don't care if a function uses the def internally
-        const defun_ranges = new Map();
+    // No defuns use outside constants
+    if (!defun_enclosing_defs_map.size) {
+        return;
+    }
 
-        let tw;
-        parent.walk((tw = new TreeWalker((node, descend) => {
-            if (node instanceof AST_Defun && defuns_of_interest.has(node)) {
+    // `defun` id to array of enclsoing `defun` id
+    // reverse of `defun_dependents_map`, used to speed up updating `defun_first_read_map`
+    const enclosing_defuns_map = new Map();
+    // temp work queue
+    const defun_queue = Array.from(defun_enclosing_defs_map.keys());
+    // `defun` ids to be tracked for definition location
+    const defuns_of_interest = new Set(defun_queue);
+    while (defun_queue.length) {
+        const defun_id = defun_queue.pop();
+        symbols_of_interest.add(defun_id);
+
+        for (const dependent of defun_dependents_map.get(defun_id) || []) {
+            if (!defuns_of_interest.has(dependent)) {
+                defuns_of_interest.add(dependent);
+                defun_queue.push(dependent);
+
+                if (!enclosing_defuns_map.has(dependent)) {
+                    enclosing_defuns_map.set(dependent, []);
+                }
+                enclosing_defuns_map.get(dependent).push(defun_id);
+            }
+        }
+    }
+
+    // All "symbols of interest", that is, defuns or defs, that we found.
+    // These are placed in order so we can check which is after which.
+    const found_symbols = [];
+    // Indices of `found_symbols` which are writes
+    const found_symbol_writes = new Set();
+    // Defun ranges are recorded because we don't care if a function uses the def internally
+    const defun_ranges = new Map();
+
+    let tw;
+    parent.walk((tw = new TreeWalker((node, descend) => {
+        if (node instanceof AST_Defun) {
+            const id = node.name.definition().id;
+            if (defuns_of_interest.has(id)) {
                 const start = found_symbols.length;
                 descend();
                 const end = found_symbols.length;
 
-                defun_ranges.set(node, { start, end });
+                defun_ranges.set(id, { start, end });
                 return true;
             }
-            // if we found a defun on the list, mark IN_DEFUN=id and descend
+        }
+        // if we found a defun on the list, mark IN_DEFUN=id and descend
 
-            if (node instanceof AST_Symbol && node.thedef) {
-                const id = node.definition().id;
-                if (symbols_of_interest.has(id)) {
-                    if (node instanceof AST_SymbolDeclaration || is_lhs(node, tw)) {
-                        found_symbol_writes.add(found_symbols.length);
-                    }
-                    found_symbols.push(id);
+        if (node instanceof AST_Symbol && node.thedef) {
+            const id = node.definition().id;
+            if (symbols_of_interest.has(id)) {
+                if (node instanceof AST_SymbolDeclaration || is_lhs(node, tw)) {
+                    found_symbol_writes.add(found_symbols.length);
                 }
+                found_symbols.push(id);
             }
-        })));
+        }
+    })));
 
-        for (const { def, defun, fname_def } of potential_conflicts) {
-            const defun_range = defun_ranges.get(defun);
+    const defun_first_read_map = new Map();
+    for (const defun of defuns_of_interest) {
+        const defun_range = defun_ranges.get(defun);
+        const defun_first_read = found_symbols.findIndex(
+            (sym, index) =>
+                sym === defun &&
+                (index < defun_range.start || index > defun_range.end)
+        );
+        if (defun_first_read === -1) {
+            continue;
+        }
+        defun_first_read_map.set(defun, defun_first_read);
 
-            // find the index in `found_symbols`, with some special rules:
-            const find = (sym_id, starting_at = 0, must_be_write = false) => {
-                let index = starting_at;
+        // Update all depdencies of `defun`
+        const queue = enclosing_defuns_map.get(defun)?.slice() ?? [];
+        const visited = new Set();
+        while (queue.length) {
+            const enclosed_defun = queue.pop();
+            // bail out on circular dependencies
+            if (visited.has(enclosed_defun)) {
+                continue;
+            }
+            visited.add(enclosed_defun);
 
-                for (;;) {
-                    index = found_symbols.indexOf(sym_id, index);
+            const enclosed_defun_first_read = defun_first_read_map.get(enclosed_defun) ?? Infinity;
+            if (enclosed_defun_first_read < defun_first_read) {
+                continue;
+            }
 
-                    if (index === -1) {
-                        break;
-                    } else if (index >= defun_range.start && index < defun_range.end) {
-                        index = defun_range.end;
-                        continue;
-                    } else if (must_be_write && !found_symbol_writes.has(index)) {
-                        index++;
-                        continue;
-                    } else {
-                        break;
-                    }
-                }
+            defun_first_read_map.set(enclosed_defun, defun_first_read);
 
-                return index;
-            };
+            for (const enclosed_enclosed_defun of enclosing_defuns_map.get(enclosed_defun) ?? []) {
+                queue.push(enclosed_enclosed_defun);
+            }
+        }
+    }
 
-            const read_defun_at = find(fname_def.id);
-            const wrote_def_at = find(def.id, read_defun_at + 1, true);
+    for (const [defun, defs] of defun_enclosing_defs_map) {
+        const defun_first_read = defun_first_read_map.get(defun);
+        if (defun_first_read === undefined) {
+            continue;
+        }
 
-            const wrote_def_after_reading_defun = read_defun_at != -1 && wrote_def_at != -1 && wrote_def_at > read_defun_at;
+        for (const def of defs) {
+            if (def.fixed === false) {
+                continue;
+            }
 
-            if (wrote_def_after_reading_defun) {
+            const def_last_write = found_symbols.findLastIndex(
+                (sym, index) => sym === def.id && found_symbol_writes.has(index)
+            );
+
+            if (defun_first_read < def_last_write) {
                 def.fixed = false;
             }
         }