--- a/lib/compress/reduce-vars.js
+++ b/lib/compress/reduce-vars.js
@@ -630,7 +630,7 @@ function handle_defined_after_hoist(parent) {
         const defun_first_read = found_symbols.findIndex(
             (sym, index) =>
                 sym === defun &&
-                (index < defun_range.start || index > defun_range.end)
+                (defun_range.start === defun_range.end || index < defun_range.start || index >= defun_range.end)
         );
         if (defun_first_read === -1) {
             continue;
@@ -663,6 +663,7 @@ function handle_defined_after_hoist(parent) {
         }
     }
 
+    // ensure write-then-read order, otherwise clear `fixed`
     for (const [defun, defs] of defun_enclosing_defs_map) {
         const defun_first_read = defun_first_read_map.get(defun);
         if (defun_first_read === undefined) {