--- a/lib/compress/reduce-vars.js
+++ b/lib/compress/reduce-vars.js
@@ -650,8 +650,7 @@ function handle_defined_after_hoist(parent) {
             visited.add(enclosed_defun);
 
             let enclosed_defun_first_read = defun_first_read_map.get(enclosed_defun);
-            if (enclosed_defun_first_read == null) enclosed_defun_first_read = Infinity;
-            if (enclosed_defun_first_read < defun_first_read) {
+            if (enclosed_defun_first_read != null && enclosed_defun_first_read < defun_first_read) {
                 continue;
             }