--- a/lib/compress/reduce-vars.js
+++ b/lib/compress/reduce-vars.js
@@ -496,7 +496,24 @@ function mark_lambda(tw, descend, compressor) {
  *   // use defined_after
  * }
  *
- * This function is called on the parent to handle this issue.
+ * Or even indirectly:
+ *
+ * B();
+ * var defined_after = true;
+ * function A() {
+ *   // use defined_after
+ * }
+ * function B() {
+ *   A();
+ * }
+ *
+ * Access a variable before declaration will either throw a ReferenceError
+ * (if the variable is declared with `let` or `const`),
+ * or get an `undefined` (if the variable is declared with `var`).
+ *
+ * If the variable is inlined into the function, the behavior will change.
+ *
+ * This function is called on the parent to disallow inlining of such variables,
  */
 function handle_defined_after_hoist(parent) {
     const defuns = [];
@@ -508,16 +525,6 @@ function handle_defined_after_hoist(parent) {
             || node instanceof AST_SimpleStatement
         ) return true;
     });
-
-    // If a function `A` uses an outside variable `B`, calling `A` before `B` is defined
-    // will either cause a ReferenceError (if `B` is declared with `let` or `const`) or
-    // get an `undefined` (if `B` is declared with `var`).
-    //
-    // We want to inline `B` into `A` if `B` is a constant and only used in `A`, but we can't
-    // do that if the above situation happens, as it would change the behavior.
-    //
-    // This function checks for such use-before-define situations, including the case where
-    // `A` is indirectly called by another function `C` that is called before `B` is defined.
 
     // `defun` id to array of `defun` that use the function
     const defun_dependents_map = new Map();