--- a/lib/compress/reduce-vars.js
+++ b/lib/compress/reduce-vars.js
@@ -638,7 +638,8 @@ function handle_defined_after_hoist(parent) {
         defun_first_read_map.set(defun, defun_first_read);
 
         // Update all depdencies of `defun`
-        const queue = enclosing_defuns_map.get(defun)?.slice() ?? [];
+        let queue = enclosing_defuns_map.get(defun);
+        queue = queue ? queue.slice() : [];
         const visited = new Set();
         while (queue.length) {
             const enclosed_defun = queue.pop();
@@ -672,9 +673,12 @@ function handle_defined_after_hoist(parent) {
                 continue;
             }
 
-            const def_last_write = found_symbols.findLastIndex(
-                (sym, index) => sym === def.id && found_symbol_writes.has(index)
-            );
+            let def_last_write = found_symbols.length;
+            while (--def_last_write >= 0) {
+                if (found_symbols[def_last_write] === def.id && found_symbol_writes.has(def_last_write)) {
+                    break;
+                }
+            }
 
             if (defun_first_read < def_last_write) {
                 def.fixed = false;