--- a/lib/compress/reduce-vars.js
+++ b/lib/compress/reduce-vars.js
@@ -649,14 +649,15 @@ function handle_defined_after_hoist(parent) {
             }
             visited.add(enclosed_defun);
 
-            const enclosed_defun_first_read = defun_first_read_map.get(enclosed_defun) ?? Infinity;
+            let enclosed_defun_first_read = defun_first_read_map.get(enclosed_defun);
+            if (enclosed_defun_first_read == null) enclosed_defun_first_read = Infinity;
             if (enclosed_defun_first_read < defun_first_read) {
                 continue;
             }
 
             defun_first_read_map.set(enclosed_defun, defun_first_read);
 
-            for (const enclosed_enclosed_defun of enclosing_defuns_map.get(enclosed_defun) ?? []) {
+            for (const enclosed_enclosed_defun of enclosing_defuns_map.get(enclosed_defun) || []) {
                 queue.push(enclosed_enclosed_defun);
             }
         }