--- a/lib/compress/reduce-vars.js
+++ b/lib/compress/reduce-vars.js
@@ -576,18 +576,14 @@ function handle_defined_after_hoist(parent) {
     // `defun` id to array of enclsoing `defun` id
     // reverse of `defun_dependents_map`, used to speed up updating `defun_first_read_map`
     const enclosing_defuns_map = new Map();
-    // temp work queue
-    const defun_queue = Array.from(defun_enclosing_defs_map.keys());
     // `defun` ids to be tracked for definition location
-    const defuns_of_interest = new Set(defun_queue);
-    while (defun_queue.length) {
-        const defun_id = defun_queue.pop();
+    const defuns_of_interest = new Set(defun_enclosing_defs_map.keys());
+    for (const defun_id of defuns_of_interest) {
         symbols_of_interest.add(defun_id);
 
         for (const dependent of defun_dependents_map.get(defun_id) || []) {
             if (!defuns_of_interest.has(dependent)) {
                 defuns_of_interest.add(dependent);
-                defun_queue.push(dependent);
 
                 if (!enclosing_defuns_map.has(dependent)) {
                     enclosing_defuns_map.set(dependent, []);
@@ -600,8 +596,8 @@ function handle_defined_after_hoist(parent) {
     // All "symbols of interest", that is, defuns or defs, that we found.
     // These are placed in order so we can check which is after which.
     const found_symbols = [];
-    // Indices of `found_symbols` which are writes
-    const found_symbol_writes = new Set();
+    // Map a symbol ID to its last write (a `found_symbols` index)
+    const symbol_last_write_map = new Map();
     // Defun ranges are recorded because we don't care if a function uses the def internally
     const defun_ranges = new Map();
 
@@ -624,7 +620,8 @@ function handle_defined_after_hoist(parent) {
             const id = node.definition().id;
             if (symbols_of_interest.has(id)) {
                 if (node instanceof AST_SymbolDeclaration || is_lhs(node, tw)) {
-                    found_symbol_writes.add(found_symbols.length);
+                    // `found_symbols.length` only becomes larger
+                    symbol_last_write_map.set(id, found_symbols.length);
                 }
                 found_symbols.push(id);
             }
@@ -645,17 +642,8 @@ function handle_defined_after_hoist(parent) {
         defun_first_read_map.set(defun, defun_first_read);
 
         // Update all depdencies of `defun`
-        let queue = enclosing_defuns_map.get(defun);
-        queue = queue ? queue.slice() : [];
-        const visited = new Set();
-        while (queue.length) {
-            const enclosed_defun = queue.pop();
-            // bail out on circular dependencies
-            if (visited.has(enclosed_defun)) {
-                continue;
-            }
-            visited.add(enclosed_defun);
-
+        const queue = new Set(enclosing_defuns_map.get(defun));
+        for (const enclosed_defun of queue) {
             let enclosed_defun_first_read = defun_first_read_map.get(enclosed_defun);
             if (enclosed_defun_first_read != null && enclosed_defun_first_read < defun_first_read) {
                 continue;
@@ -664,12 +652,13 @@ function handle_defined_after_hoist(parent) {
             defun_first_read_map.set(enclosed_defun, defun_first_read);
 
             for (const enclosed_enclosed_defun of enclosing_defuns_map.get(enclosed_defun) || []) {
-                queue.push(enclosed_enclosed_defun);
+                queue.add(enclosed_enclosed_defun);
             }
         }
     }
 
     // ensure write-then-read order, otherwise clear `fixed`
+    // This is safe because last-writes (found_symbol_writes) are assumed to be as late as possible, and first-reads (defun_first_read_map) are assumed to be as early as possible.
     for (const [defun, defs] of defun_enclosing_defs_map) {
         const defun_first_read = defun_first_read_map.get(defun);
         if (defun_first_read === undefined) {
@@ -681,12 +670,7 @@ function handle_defined_after_hoist(parent) {
                 continue;
             }
 
-            let def_last_write = found_symbols.length;
-            while (--def_last_write >= 0) {
-                if (found_symbols[def_last_write] === def.id && found_symbol_writes.has(def_last_write)) {
-                    break;
-                }
-            }
+            let def_last_write = symbol_last_write_map.get(def.id) || 0;
 
             if (defun_first_read < def_last_write) {
                 def.fixed = false;