--- a/lib/compress/reduce-vars.js
+++ b/lib/compress/reduce-vars.js
@@ -94,8 +94,7 @@ import {
 
     walk,
     walk_body,
-
-    TreeWalker,
+    walk_parent,
 } from "../ast.js";
 import { HOP, make_node, noop } from "../utils/index.js";
 
@@ -519,19 +518,23 @@ function handle_defined_after_hoist(parent) {
     const defuns = [];
     walk(parent, node => {
         if (node === parent) return;
-        if (node instanceof AST_Defun) defuns.push(node);
+        if (node instanceof AST_Defun) {
+            defuns.push(node);
+            return true;
+        }
         if (
             node instanceof AST_Scope
             || node instanceof AST_SimpleStatement
         ) return true;
     });
 
-    // `defun` id to array of `defun` that use the function
-    const defun_dependents_map = new Map();
+    // `defun` id to array of `defun` it uses
+    const enclosing_defuns_map = new Map();
     // `defun` id to array of enclosing `def` that are used by the function
     const defun_enclosing_defs_map = new Map();
     // all symbol ids that will be tracked for read/write
     const symbols_of_interest = new Set();
+    const defuns_of_interest = new Set();
 
     for (const defun of defuns) {
         const fname_def = defun.name.definition();
@@ -546,25 +549,32 @@ function handle_defined_after_hoist(parent) {
                 continue;
             }
 
+            symbols_of_interest.add(def.id);
+
             // found a reference to another function
             if (
                 def.assignments === 0
                 && def.orig.length === 1
                 && def.orig[0] instanceof AST_SymbolDefun
             ) {
-                if (!defun_dependents_map.has(def.id)) {
-                    defun_dependents_map.set(def.id, []);
+                defuns_of_interest.add(fname_def.id);
+                symbols_of_interest.add(fname_def.id);
+
+                if (!enclosing_defuns_map.has(fname_def.id)) {
+                    enclosing_defuns_map.set(fname_def.id, []);
                 }
-                defun_dependents_map.get(def.id).push(fname_def.id);
+                enclosing_defuns_map.get(fname_def.id).push(def.id);
+
                 continue;
             }
 
             enclosing_defs.push(def);
-            symbols_of_interest.add(def.id);
         }
 
         if (enclosing_defs.length) {
             defun_enclosing_defs_map.set(fname_def.id, enclosing_defs);
+            defuns_of_interest.add(fname_def.id);
+            symbols_of_interest.add(fname_def.id);
         }
     }
 
@@ -573,74 +583,38 @@ function handle_defined_after_hoist(parent) {
         return;
     }
 
-    // `defun` id to array of enclsoing `defun` id
-    // reverse of `defun_dependents_map`, used to speed up updating `defun_first_read_map`
-    const enclosing_defuns_map = new Map();
-    // `defun` ids to be tracked for definition location
-    const defuns_of_interest = new Set(defun_enclosing_defs_map.keys());
-    for (const defun_id of defuns_of_interest) {
-        symbols_of_interest.add(defun_id);
-
-        for (const dependent of defun_dependents_map.get(defun_id) || []) {
-            if (!defuns_of_interest.has(dependent)) {
-                defuns_of_interest.add(dependent);
-
-                if (!enclosing_defuns_map.has(dependent)) {
-                    enclosing_defuns_map.set(dependent, []);
-                }
-                enclosing_defuns_map.get(dependent).push(defun_id);
-            }
-        }
-    }
-
-    // All "symbols of interest", that is, defuns or defs, that we found.
-    // These are placed in order so we can check which is after which.
-    const found_symbols = [];
-    // Map a symbol ID to its last write (a `found_symbols` index)
+    // Increment to count "symbols of interest" (defuns or defs) that we found.
+    // These are tracked in AST order so we can check which is after which.
+    let symbol_index = 1;
+    // Map a defun ID to its first read (a `symbol_index`)
+    const defun_first_read_map = new Map();
+    // Map a symbol ID to its last write (a `symbol_index`)
     const symbol_last_write_map = new Map();
-    // Defun ranges are recorded because we don't care if a function uses the def internally
-    const defun_ranges = new Map();
 
-    let tw;
-    parent.walk((tw = new TreeWalker((node, descend) => {
-        if (node instanceof AST_Defun) {
-            const id = node.name.definition().id;
-            if (defuns_of_interest.has(id)) {
-                const start = found_symbols.length;
-                descend();
-                const end = found_symbols.length;
-
-                defun_ranges.set(id, { start, end });
-                return true;
-            }
-        }
-        // if we found a defun on the list, mark IN_DEFUN=id and descend
-
+    walk_parent(parent, (node, walk_info) => {
         if (node instanceof AST_Symbol && node.thedef) {
             const id = node.definition().id;
+
+            symbol_index++;
+
+            // Track last-writes to symbols
             if (symbols_of_interest.has(id)) {
-                if (node instanceof AST_SymbolDeclaration || is_lhs(node, tw)) {
-                    // `found_symbols.length` only becomes larger
-                    symbol_last_write_map.set(id, found_symbols.length);
+                if (node instanceof AST_SymbolDeclaration || is_lhs(node, walk_info.parent())) {
+                    symbol_last_write_map.set(id, symbol_index);
                 }
-                found_symbols.push(id);
+            }
+
+            // Track first-reads of defuns (refined later)
+            if (defuns_of_interest.has(id)) {
+                if (!defun_first_read_map.has(id) && !is_recursive_ref(walk_info, id)) {
+                    defun_first_read_map.set(id, symbol_index);
+                }
             }
         }
-    })));
+    });
 
-    const defun_first_read_map = new Map();
-    for (const defun of defuns_of_interest) {
-        const defun_range = defun_ranges.get(defun);
-        const defun_first_read = found_symbols.findIndex(
-            (sym, index) =>
-                sym === defun &&
-                (defun_range.start === defun_range.end || index < defun_range.start || index >= defun_range.end)
-        );
-        if (defun_first_read === -1) {
-            continue;
-        }
-        defun_first_read_map.set(defun, defun_first_read);
-
+    // Refine `defun_first_read_map` to be as high as possible
+    for (const [defun, defun_first_read] of defun_first_read_map) {
         // Update all depdencies of `defun`
         const queue = new Set(enclosing_defuns_map.get(defun));
         for (const enclosed_defun of queue) {