… test/compress/issue-1429.js
… test/compress/reduce_vars.js