Fixes #1429 

This allows variables to be reduced when used in a function that has indirect calls.

I'm new to terser so it might have bugs.

Benchmark result:

| url                                                                                | input size | before output size | after output size | size diff     | before time | after time | time diff |
| ---------------------------------------------------------------------------------- | ---------- | ------------------ | ----------------- | ------------- | ----------- | ---------- | --------- |
| https://code.jquery.com/jquery-3.2.1.js                                            | 268039     | 86877              | 86877             | 0             | 674.72494   | 672.01946  | -0.40%    |
| https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.6.4/angular.js                 | 1249863    | 174525             | 174482            | -43 (-0.02%)  | 1175.19904  | 1187.24018 | +1.02%    |
| https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js                        | 1590107    | 466647             | 466637            | -10 (-0.00%)  | 2040.97124  | 1998.89482 | -2.06%    |
| https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js                    | 69707      | 37046              | 37046             | 0             | 326.45528   | 319.89132  | -2.01%    |
| https://unpkg.com/react@15.3.2/dist/react.js                                       | 701412     | 205980             | 205959            | -21 (-0.01%)  | 1035.3006   | 1031.51174 | -0.37%    |
| http://builds.emberjs.com/tags/v2.11.0/ember.prod.js                               | 1852178    | 528006             | 528006            | 0             | 1900.31164  | 1906.68276 | +0.34%    |
| https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js                                   | 539590     | 70560              | 70560             | 0             | 712.81798   | 717.3766   | +0.64%    |
| https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js                              | 451131     | 212579             | 212498            | -81 (-0.04%)  | 1314.11204  | 1327.10142 | +0.99%    |
| https://raw.githubusercontent.com/kangax/html-minifier/v3.5.7/dist/htmlminifier.js | 1063075    | 511806             | 511105            | -701 (-0.14%) | 2020.1799   | 1999.70882 | -1.01%    |

In my real project it can shave another 1% (20KB).