@fabiosantoscode 2024-08-07
This looks really promising! Node 10 isn't liking the `?.` syntax though, and I guess .findLastIndex is also out of the table.

I'll have a closer look this afternoon.