@fabiosantoscode 2024-08-07
I finally understood the whole thing :) I see that you're using indices in AST traversal order, and then comparing the first read of a defun to the last write of its dependency.

This would not be enough (since AST order is not the same as program order), but the first read of the defun is changed so as to become as small as possible according to its dependents. This accepts false negatives, but it should have no false positives.