@yume-chan 2024-08-11
> I see that you're using indices in AST traversal order, and then comparing the first read of a defun to the last write of its dependency.

The original logic already checks the first read of a defun (thus its containing variables) and the last write of those variables to determine if it's safe to inline those variables. I extended the logic to include indirect calls to enable more inlining opportunities (by propagating the read location of a defun to all its dependencies. If all reads happen after the last write, it should still be safe to inline).

> This would not be enough (since AST order is not the same as program order)

Do you mean there are cases where my version may incorrectly inline a variable and change the program behavior?

> This accepts false negatives, but it should have no false positives.

Yes, it's on the safe side.