@fabiosantoscode 2024-08-11
No it's all good. I'll try to merge it on Monday.

> The original logic already checks the first read of a defun (thus its containing variables) and the last write of those variables to determine if it's safe to inline those variables

I'll add a comment to explain this, because I forgot what I did originally and it took me a while.

I'll also change the `while (queue.length)` style of traversal to the following style to avoid creating a separate variable. I *think* it won't be slower but I'll have to check.

```js
const queue = new Set(initial) // iterates in insertion order
for (const item of queue) {
  // Enqueue with queue.add(), no-op if already added
}
```