@fabiosantoscode 2024-08-13
I force-pushed because I did a rebase from the main branch. I'm currently looking at the first-read calculations, which I believe can have the last-index calculation removed or greatly simplified.